Add files via upload
The Snake Game is a classic arcade game where the player controls a snake on the screen. The snake moves around, eats food, and grows longer with each meal. The objective is to grow the snake as long as possible without running into the walls or colliding with itself. The game increases in difficulty as the snake becomes longer and space on the screen decreases.